Changes to models
- Changed models and implementation strategy
