+++ src/models/queue_message_content.py
@@ -6,110 +6,6 @@ class ValidationResult:
     validation_message: str = ''
 
 
-class Upload:
-
-    def __init__(self, data: dict):
-        upload_data = data.get('data', None)
-        self._message = data.get('message', None)
-        self._message_type = data.get('messageType', None)
-        self._message_id = data.get('messageId', '')
-        self._published_date = data.get('publishedDate', None)
-        self.data = UploadData(data=upload_data) if upload_data else {}
-
-    @property
-    def message(self):
-        return self._message
-
-    @message.setter
-    def message(self, value):
-        self._message = value
-
-    @property
-    def message_type(self):
-        return self._message_type
-
-    @message_type.setter
-    def message_type(self, value):
-        self._message_type = value
-
-    @property
-    def message_id(self):
-        return self._message_id
-
-    @message_id.setter
-    def message_id(self, value):
-        self._message_id = value
-
-    @property
-    def published_date(self):
-        return self._published_date
-
-    @published_date.setter
-    def published_date(self, value):
-        self._published_date = value
-
-    def to_json(self):
-        self.data = self.data.to_json()
-        return to_json(self.__dict__)
-
-    def data_from(self):
-        message = self
-        if isinstance(message, str):
-            message = json.loads(self)
-        if message:
-            try:
-                return Upload(data=message)
-            except Exception as e:
-                error = str(e).replace('Upload', 'Invalid parameter,')
-                raise TypeError(error)
-
-
-class UploadData:
-    def __init__(self, data: dict):
-        polygon = data.get('polygon', None)
-        request = data.get('request', None)
-        meta = data.get('meta', None)
-
-        response = data.get('response', None)
-        self._stage = data.get('stage', '')
-        self.request = Request(data=request) if request else {}
-        self.meta = Meta(data=meta) if meta else {}
-        self.response = Response(data=response) if response else {}
-        self._tdei_record_id = data.get('tdei_record_id', '')
-        self._tdei_project_group_id = data.get('tdei_project_group_id', '')
-        self._user_id = data.get('user_id', '')
-
-    @property
-    def stage(self): return self._stage
-
-    @stage.setter
-    def stage(self, value): self._stage = value
-
-    @property
-    def tdei_record_id(self): return self._tdei_record_id
-
-    @tdei_record_id.setter
-    def tdei_record_id(self, value): self._tdei_record_id = value
-
-    @property
-    def tdei_project_group_id(self): return self._tdei_project_group_id
-
-    @tdei_project_group_id.setter
-    def tdei_project_group_id(self, value): self._tdei_project_group_id = value
-
-    @property
-    def user_id(self): return self._user_id
-
-    @user_id.setter
-    def user_id(self, value): self._user_id = value
-
-    def to_json(self):
-        self.request = to_json(self.request.__dict__)
-        self.meta = to_json(self.meta.__dict__)
-        self.response = to_json(self.response.__dict__)
-        return to_json(self.__dict__)
-
-
 class Request:
     def __init__(self, data: dict):
         self._tdei_project_group_id = data.get('tdei_project_group_id', '')
