+++ src/osw_format.py
@@ -6,14 +6,9 @@ from pathlib import Path
 from .config import Settings
 from osm_osw_reformatter import Formatter
 
-ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
-# Path used for download file generation.
-DOWNLOAD_FILE_PATH = f'{Path.cwd()}/downloads'
-
 logging.basicConfig()
-logger = logging.getLogger('FORMAT')
+logger = logging.getLogger('osw-formatter')
 logger.setLevel(logging.INFO)
-
 
 class OSWFormat:
     def __init__(self, file_path=None, storage_client=None, prefix=None):
@@ -27,6 +22,7 @@ class OSWFormat:
         self.file_relative_path = file_path.split('/')[-1]
         self.client = self.storage_client.get_container(container_name=self.container_name)
         self.prefix = prefix
+        self.download_dir = settings.get_download_directory()
 
     def format(self):
         root, ext = os.path.splitext(self.file_relative_path)
@@ -35,10 +31,10 @@ class OSWFormat:
 
             try:
                 logger.info(f' Downloaded file path: {downloaded_file_path}')
-                f = Formatter(workdir=DOWNLOAD_FILE_PATH, file_path=downloaded_file_path, prefix=self.prefix)
-                formatter = f.osw2osm()
+                formatter = Formatter(workdir=self.download_dir, file_path=downloaded_file_path, prefix=self.prefix)
+                formatter_response = formatter.osw2osm()
                 OSWFormat.clean_up(downloaded_file_path)
-                return formatter
+                return formatter_response
             except Exception as err:
                 traceback.print_exc()
                 logger.error(f' Error While Formatting File: {str(err)}')
@@ -53,10 +49,10 @@ class OSWFormat:
         try:
             if file.file_path:
                 file_path = os.path.basename(file.file_path)
-                with open(f'{DOWNLOAD_FILE_PATH}/{file_path}', 'wb') as blob:
+                with open(f'{self.download_dir}/{file_path}', 'wb') as blob:
                     blob.write(file.get_stream())
-                logger.info(f' File downloaded to location: {DOWNLOAD_FILE_PATH}/{file_path}')
-                return f'{DOWNLOAD_FILE_PATH}/{file_path}'
+                logger.info(f' File downloaded to location: {self.download_dir}/{file_path}')
+                return f'{self.download_dir}/{file_path}'
             else:
                 logger.info(' File not found!')
         except Exception as e:
@@ -69,6 +65,6 @@ class OSWFormat:
             logger.info(f' Removing File: {path}')
             os.remove(path)
         else:
-            folder = os.path.join(DOWNLOAD_FILE_PATH, path)
+            folder = os.path.join(self.download_dir, path)
             logger.info(f' Removing Folder: {folder}')
             shutil.rmtree(folder, ignore_errors=True)
